LF-4699 Improve LineChart * extract functions * adjust xAxis padding
